fix(store/commitment/iavl): honor tree.Remove error firstly
Previously the check was if !res and then it would return
the error below but that has a consequence of ignoring/swallowing
up the error returned by (*iavl.MutableTree).Remove. Even more is
the fact that when !res, an error occurred and it could be anything
else, thus checking against the error firstly ensures that we only
return a synthesized error when we are certain the key wasn't in the tree.

Fixes #18650